+++ R/calcWINS_hce.R
@@ -16,16 +16,19 @@
 #' @md
 #' @seealso [hce::calcWINS()], [hce::calcWINS.formula()], [hce::calcWINS.data.frame()].
 #' @references
-#' The theory of win statistics is covered in the following papers. 
-#' * For the win proportion CI calculation see 
+#' The theory of win statistics is covered in the following papers: 
+#' * Win proportion and win odds confidence interval calculation: 
+#' \cr \cr Bamber D (1975) "The area above the ordinal dominance graph and the area below the receiver operating characteristic graph." Journal of Mathematical Psychology 12.4: 387-415. <doi:10.1016/0022-2496(75)90001-2>.
+#' \cr \cr DeLong ER et al. (1988) "Comparing the Areas Under Two or More Correlated Receiver Operating Characteristic Curves: A Nonparametric Approach." Biometrics 44.3: 837-845. <doi:10.2307/2531595>.
+#' \cr \cr Brunner E et al. (2021) "Win odds: an adaptation of the win ratio to include ties." Statistics in Medicine 40.14: 3367-3384. <doi:10.1002/sim.8967>.
 #' \cr \cr Gasparyan SB et al. (2021) "Adjusted win ratio with stratification: calculation methods and interpretation." Statistical Methods in Medical Research 30.2: 580-611. <doi:10.1177/0962280220942558>.
-#' * The win odds CI is calculated using the formula in 
 #' \cr \cr Gasparyan SB et al. (2021) "Power and sample size calculation for the win odds test: application to an ordinal endpoint in COVID-19 trials." Journal of Biopharmaceutical Statistics 31.6: 765-787. <doi:10.1080/10543406.2021.1968893>.
-#' * The win ratio the first CI uses the standard error derived from the standard error of the `gamma` statistic presented in
-#' \cr \cr Gasparyan SB, Kowalewski EK, Buenconsejo J, Koch GG. (2023) “Hierarchical Composite Endpoints in COVID-19: The DARE-19 Trial.” In Case Studies in Innovative Clinical Trials, Chapter 7, 95–148. Chapman; Hall/CRC. <doi:10.1201/9781003288640-7>.
-#' * The win ratio the second CI uses the standard error presented in
+#' \cr \cr Brunner E, Konietschke F. (2025) "An unbiased rank-based estimator of the Mann–Whitney variance including the case of ties." Statistical Papers 66.20. <doi:10.1007/s00362-024-01635-0>.
+#' * Win ratio: the first CI utilizes the standard error derived from the `gamma` statistic standard error as outlined by:
+#' \cr \cr Gasparyan SB, Kowalewski EK, Buenconsejo J, Koch GG. (2023) "Hierarchical Composite Endpoints in COVID-19: The DARE-19 Trial." In Case Studies in Innovative Clinical Trials, Chapter 7, 95–148. Chapman; Hall/CRC. <doi:10.1201/9781003288640-7>.
+#' * Win ratio: the second CI utilizes the standard error presented by:
 #' \cr \cr Yu RX, Ganju J. (2022) "Sample size formula for a win ratio endpoint." Statistics in Medicine 41.6: 950-63. <doi:10.1002/sim.9297>.
-#' * The Goodman Kruskal's `gamma` and its CI match those in `DescTools::GoodmanKruskalGamma()` and are based on
+#' * Goodman Kruskal's `gamma` and CI: matches implementation in `DescTools::GoodmanKruskalGamma()` and based on:
 #' \cr \cr Agresti A. (2002) Categorical Data Analysis. John Wiley & Sons, pp. 57-59. <doi:10.1002/0471249688>.
 #' \cr \cr Brown MB, Benedetti JK. (1977) "Sampling Behavior of Tests for Correlation in Two-Way Contingency Tables." Journal of the American Statistical Association 72, 309-315. <doi:10.1080/01621459.1977.10480995>.
 #' \cr \cr Goodman LA, Kruskal WH. (1954) "Measures of association for cross classifications." Journal of the American Statistical Association 49, 732-764. <doi:10.1080/01621459.1954.10501231>.
@@ -37,6 +40,9 @@
 #' # Example 2
 #' COVID19bHCE <- hce(GROUP = COVID19b$GROUP, TRTP = COVID19b$TRTP)
 #' calcWINS(COVID19bHCE, ref = "Placebo", WOnull = 1.1, alpha = 0.01)
+#' # Example 3
+#' calcWINS(COVID19HCE, SE_WP_Type = "unbiased")$WP
+#' calcWINS(COVID19HCE, SE_WP_Type = "biased")$WP
 calcWINS.hce <- function(x, ...){
   Args <- base::list(...)
   x <- as_hce(x)
@@ -51,7 +57,7 @@ calcWINS.hce <- function(x, ...){
   else WOnull <- 1
   
   res <- calcWINS.data.frame(x = x, AVAL = "AVAL", TRTP = "TRTP", 
-                             ref = ref, alpha = alpha, WOnull = WOnull)
+                             ref = ref, alpha = alpha, WOnull = WOnull, SE_WP_Type = Args[["SE_WP_Type"]])
   res$ref <- base::paste(unique(x$TRTP)[unique(x$TRTP) != ref],"vs", ref)
   res$Input <- data.frame(alpha = alpha, WOnull = WOnull)
   res