+++ R/simHCE.R
@@ -12,7 +12,7 @@
 #' @param yeardays number of days in a year.
 #' @param pat scale of provided event rates (per `pat`-years).
 #' @param shape shape of the Weibull distribution for time-to-event outcomes. Default is exponential distribution with `shape = 1`.
-#' @param alpha Gumbel dependence coefficient of the Weibull distributions for time-to-event outcomes. Default is `alpha = 1` which assumes independence of time-to-event outcomes. Must be above or equal to 1.
+#' @param theta Gumbel dependence coefficient of the Weibull distributions for time-to-event outcomes. Default is `theta = 1` which assumes independence of time-to-event outcomes. Must be above or equal to 1.
 #' @param logC logical, whether to use log-normal distribution for the continuous outcome.
 #' @param seed for generating random numbers.
 #' @param dec decimal places for the continuous outcome used for rounding. The default is `dec = 2`.
@@ -49,17 +49,17 @@
 #' Rates_A <- c(10, 20)
 #' Rates_P <- c(20, 20)
 #' dat1 <- simHCE(n = 2500, TTE_A = Rates_A, TTE_P = Rates_P, 
-#' CM_A = -3, CM_P = -6, CSD_A = 15, fixedfy = 3, alpha = 1, seed = 1)
+#' CM_A = -3, CM_P = -6, CSD_A = 15, fixedfy = 3, theta = 1, seed = 1)
 #' dat2 <- simHCE(n = 2500, TTE_A = Rates_A, TTE_P = Rates_P, 
-#' CM_A = -3, CM_P = -6, CSD_A = 15, fixedfy = 3, alpha = 10, seed = 1)
+#' CM_A = -3, CM_P = -6, CSD_A = 15, fixedfy = 3, theta = 1.0001, seed = 1)
 #' calcWO(dat1)
 #' calcWO(dat2)
 
 
-simHCE <- function(n, n0 = n, TTE_A, TTE_P, CM_A, CM_P, CSD_A = 1, CSD_P = CSD_A, fixedfy = 1, yeardays = 360, pat = 100, shape = 1, alpha = 1, logC = FALSE, seed = NULL, dec = 2 ){
+simHCE <- function(n, n0 = n, TTE_A, TTE_P, CM_A, CM_P, CSD_A = 1, CSD_P = CSD_A, fixedfy = 1, yeardays = 360, pat = 100, shape = 1, theta = 1, logC = FALSE, seed = NULL, dec = 2 ){
   if(base::length(TTE_P) != base::length(TTE_A))
     stop("Event rate vectors for active and placebo groups should have the same length.")
-  
+  stopifnot("`theta` must be greater than or equal to 1." = theta >= 1)
   n <- n[1]
   n0 <- n0[1]
   CM_A <- CM_A[1]
@@ -73,7 +73,7 @@ simHCE <- function(n, n0 = n, TTE_A, TTE_P, CM_A, CM_P, CSD_A = 1, CSD_P = CSD_A
   logC <- as.logical(logC[1])
   shape <- shape[1]
   dec <- dec[1]
-  alpha <- alpha[1]
+  theta <- theta[1]
 
   fixedf <- yeardays*fixedfy
   TTE_P <- TTE_P/yeardays/pat
@@ -81,40 +81,32 @@ simHCE <- function(n, n0 = n, TTE_A, TTE_P, CM_A, CM_P, CSD_A = 1, CSD_P = CSD_A
   l <- base::length(TTE_A)
   if(is.null(seed)){
     seed <- base::as.numeric(base::Sys.time())
+    seed <- base::round(seed)
   }
   seed <- seed[1]
   base::set.seed(seed)
-  if(alpha < 1)
-    stop("Alpha should be greater or equal to 1.")
-  rgumbel <- function(n, dim = 2, alpha = 1){
-    exprand <- matrix(stats::rexp(dim * n), c(n, dim))
-    unifpirand <- stats::runif(n, 0, pi)
-    exprand2 <- stats::rexp(n)
-    beta <- 1/alpha
-    stablerand <- sin((1 - beta) * unifpirand)^((1 - beta)/beta) * 
-      (sin(beta * unifpirand))/(sin(unifpirand))^(1/beta)
-    stablerand <- stablerand/(exprand2^(alpha - 1))
-    invphi <- function (t, alpha = 1){
-      exp(-t^(1/alpha))
-    } 
-    unifrand <- invphi(exprand/stablerand, alpha)
-    return(unifrand)  
+  invW <- function(u, scale, shape) {
+    scale * (- log( 1 - u))^(1/shape)
   }
-  
-  invW <- function(u, beta, gamma) {
-    beta * (- log( 1 - u))^(1/gamma)
-  }
-  
-  if(alpha == 1){
+  if(theta == 1){
     M_P <- base::sapply(1/TTE_P, stats::rweibull, n = n0, shape = shape)
     M_A <- base::sapply(1/TTE_A, stats::rweibull, n = n, shape = shape)
-  } else if(alpha > 1){
-    Y <- rgumbel(n, alpha = alpha, dim = l)
-    X <- rgumbel(n0, alpha = alpha, dim = l)
-    Y1 <- lapply(Y, invW, beta = 1/TTE_A, gamma = shape)
-    M_A <- do.call(rbind, Y1)
-    X1 <- lapply(X, invW, beta = 1/TTE_P, gamma = shape)
-    M_P <- do.call(rbind, X1)
+  } else if(theta > 1){
+    Y <- rGumbel(n, theta = theta, dim = l)
+    X <- rGumbel(n0, theta = theta, dim = l)
+    
+    d <- data.frame(AVAL = as.numeric(Y), RATE = rep(TTE_A, each = n), TTE = rep(1:l, each = n))
+    d0 <- lapply(split(d, d$TTE), function(x) invW(x[, 1], scale = 1/x[, 2], shape = shape))
+    M_A <- do.call(cbind, d0)
+    dimnames(M_A) <- NULL
+    d <- data.frame(AVAL = as.numeric(X), RATE = rep(TTE_P, each = n0), TTE = rep(1:l, each = n0))
+    d0 <- lapply(split(d, d$TTE), function(x) invW(x[, 1], scale = 1/x[, 2], shape = shape))
+    M_P <- do.call(cbind, d0)
+    dimnames(M_P) <- NULL
+    #Y1 <- lapply(Y, invW, scale = 1/TTE_A, shape = shape)
+    #M_A <- do.call(rbind, Y1)
+    #X1 <- lapply(X, invW, scale = 1/TTE_P, shape = shape)
+    #M_P <- do.call(rbind, X1)
   }
   M_P <- round(M_P)
   M_P <- base::as.data.frame(M_P)
@@ -129,7 +121,6 @@ simHCE <- function(n, n0 = n, TTE_A, TTE_P, CM_A, CM_P, CSD_A = 1, CSD_P = CSD_A
   dat <- base::rbind(M_A, M_P)
   dat$PADY <- fixedf
   dat$seed <- seed
-  
   
   dat$AVALT <- base::apply(dat[, base::startsWith(names(dat), "TTE"), drop = FALSE], 1, function(x) x[x <= fixedf][1])
   dat$AVALT[is.na(dat$AVALT)] <- fixedf
@@ -162,5 +153,70 @@ simHCE <- function(n, n0 = n, TTE_A, TTE_P, CM_A, CM_P, CSD_A = 1, CSD_P = CSD_A
   return(dat)
 }
 
+phi <- function(t, theta = 1) {
+  exp(- t^(1/theta))
+}
 
+#' Simulate from a Gumbel copula (version 1)
+#'
+#' @param n number of observations. 
+#' @param dim dimension of the copula.
+#' @param theta coefficient of dependence. 
+#'
+#' @return a random vector.
+#' @noRd
+rgumbel <- function(n, dim = 2, theta = 1){
+  stopifnot("`theta` must be greater than or equal to 1." = theta >= 1)
+  if(length(n) > 1)
+    n <- length(n)
+  if(length(dim) > 1)
+    dim <- length(dim)
+  exprand <- matrix(stats::rexp(dim * n), c(n, dim))
+  unifpirand <- stats::runif(n, 0, pi)
+  exprand2 <- stats::rexp(n)
+  beta <- 1/theta
+  stablerand <- sin((1 - beta) * unifpirand)^((1 - beta)/beta) * 
+    (sin(beta * unifpirand))/(sin(unifpirand))^(1/beta)
+  stablerand <- stablerand/(exprand2^(theta - 1))
+  unifrand <- phi(exprand/stablerand, theta)
+  return(unifrand)  
+}
 
+#' Simulate from a Gumbel copula (version 2)
+#'
+#' @param n number of observations. 
+#' @param dim dimension of the copula.
+#' @param theta coefficient of dependence. 
+#'
+#' @return a random vector.
+#' @noRd
+rGumbel <- function(n, dim = 2, theta = 1) {
+  stopifnot("`theta` must be greater than or equal to 1." = theta >= 1)
+  if(length(dim) > 1)
+    dim <- length(dim)
+  if(length(n) > 1)
+    n <- length(n)
+  rSTBL <- function(n, theta){
+    Theta <- stats::runif(n, -pi/2, pi/2)
+    W <- stats::rexp(n)
+    beta <- 1
+    alpha <- 1/theta
+    gamma <- (cos(pi/(2*theta)))^theta
+    delta <- ifelse(theta == 1, 1, 0)
+    btan <- beta*tan(alpha*pi/2)
+    theta0 <- atan(btan)/alpha
+    Ctan <- (1 + btan^2)^(1/(2*alpha))
+    alpha0 <- alpha*(theta0 + Theta)
+    if(alpha != 1){
+      Z <- sin(alpha0)*Ctan/cos(Theta)^(1/alpha)*(cos(alpha0 - Theta)/W)^((1 - alpha)/alpha)
+    } else {
+      pb <- pi/2 + beta*Theta
+      Z <- 2/pi*(pb*tan(Theta) - beta*log(pi/2*W*cos(Theta)/pb))
+    }
+    Z * gamma + delta
+  }
+  exprand <- matrix(stats::rexp(dim * n), c(n, dim))
+  stablerand <- rSTBL(n, theta = theta)
+  unifrand <- phi(exprand/stablerand, theta)
+  return(unifrand)
+}