+++ inst/doc/hce.R
@@ -4,7 +4,7 @@ knitr::opts_chunk$set(
   comment = "#>"
 )
 
-## ----echo=FALSE, out.width = '25%'--------------------------------------------
+## ----echo=FALSE, out.width = '33%'--------------------------------------------
 knitr::include_graphics("hex-hce.png")
 
 ## ----setup--------------------------------------------------------------------
@@ -51,19 +51,23 @@ dat$GROUP <- factor(dat$GROUP, levels = Order)
 table(dat$GROUP, dat$TRTP)
 
 ## ----echo=FALSE---------------------------------------------------------------
-knitr::kable(head(ADSL), caption = "ADSL dataset structure\\label{adsl}")
-
-## ----echo=FALSE---------------------------------------------------------------
-knitr::kable(head(ADLB), caption = "ADLB dataset structure\\label{adlb}")
-
-## ----echo=FALSE---------------------------------------------------------------
-knitr::kable(head(ADET), caption = "ADET dataset structure\\label{adet}")
-
-## ----echo=FALSE---------------------------------------------------------------
 HCE3 <- data.frame(Order = c("I", "II", "III", "IV"), 
                   Category = c("Cardiovascular death", 
                                "Total (first and recurrent) HF hospitalizations", 
                                "Total urgent HF visits", 
                                "Improvement/deterioration in KCCQ-TSS"))
 HCE3
+
+## -----------------------------------------------------------------------------
+Rates_A <- c(10, 20)
+Rates_P <- c(20, 20)
+dat1 <- simHCE(n = 2500, TTE_A = Rates_A, TTE_P = Rates_P, 
+CM_A = -3, CM_P = -6, CSD_A = 15, fixedfy = 3, theta = 1, seed = 1)
+dat2 <- simHCE(n = 2500, TTE_A = Rates_A, TTE_P = Rates_P, 
+CM_A = -3, CM_P = -6, CSD_A = 15, fixedfy = 3, theta = 1.0001, seed = 1)
+dat3 <- simHCE(n = 2500, TTE_A = Rates_A, TTE_P = Rates_P, 
+CM_A = -3, CM_P = -6, CSD_A = 15, fixedfy = 3, theta = 10, seed = 1)
+calcWO(dat1)
+calcWO(dat2)
+calcWO(dat3)